refactor: split up some validation functions
